+++ lib/src/api/requests/route_definitions.dart
@@ -27,6 +27,17 @@ class UserRoutes {
   static final Route create = Route.post('/user/register');
 }
 
+class AccountRoutes {
+  const AccountRoutes._();
+
+  static final Route getAll = Route.get('/account');
+  static final Route getById = Route.get('/account/{accountId}');
+  static final Route deleteById = Route.delete('/account/{accountId}');
+  static final Route patchById = Route.patch('/account/{accountId}');
+  static final Route getTransactions = Route.get('/account/{accountId}/transactions');
+  static final Route create = Route.post('/account');
+}
+
 class CurrencyRoutes {
   const CurrencyRoutes._();
 
@@ -34,5 +45,15 @@ class CurrencyRoutes {
   static final Route create = Route.post('/currency');
   static final Route getById = Route.get('/currency/{currencyId}');
   static final Route deleteById = Route.delete('/currency/{currencyId}');
-  static final Route updateById = Route.patch('/currency/{currencyId}');
+  static final Route patchById = Route.patch('/currency/{currencyId}');
+}
+
+class TransactionRoutes {
+  const TransactionRoutes._();
+
+  static final Route getAll = Route.get('/transaction');
+  static final Route getById = Route.get('/transaction/{transactionId}');
+  static final Route deleteById = Route.delete('/transaction/{transactionId}');
+  static final Route patchById = Route.patch('/transaction/{transactionId}');
+  static final Route create = Route.post('/transaction');
 }