+++ lib/src/api/restrr.dart
@@ -60,12 +60,41 @@ abstract class Restrr {
 
   Future<bool> deleteAllSessions();
 
+  /* Accounts */
+
+  Future<Account> createAccount(
+      {required String name, required int originalBalance, required Id currencyId, String? description, String? iban});
+
+  List<Account> getAccounts();
+
+  Future<Account> retrieveAccountById(Id id, {bool forceRetrieve = false});
+
+  Future<Paginated<Account>> retrieveAllAccounts({int page = 1, int limit = 25, bool forceRetrieve = false});
+
   /* Currencies */
 
   Future<Currency> createCurrency(
       {required String name, required String symbol, required int decimalPlaces, String? isoCode});
 
+  List<Currency> getCurrencies();
+
   Future<Currency> retrieveCurrencyById(Id id, {bool forceRetrieve = false});
 
   Future<Paginated<Currency>> retrieveAllCurrencies({int page = 1, int limit = 25, bool forceRetrieve = false});
+
+  /* Transactions */
+
+  Future<Transaction> createTransaction(
+      {required int amount,
+      required Id currencyId,
+      required DateTime executedAt,
+      required String name,
+      String? description,
+      Id? sourceId,
+      Id? destinationId,
+      Id? budgetId});
+
+  Future<Transaction> retrieveTransactionById(Id id, {bool forceRetrieve = false});
+
+  Future<Paginated<Transaction>> retrieveAllTransactions({int page = 1, int limit = 25, bool forceRetrieve = false});
 }