+++ lib/src/internal/utils/request_utils.dart
@@ -8,17 +8,48 @@ import '../requests/responses/rest_response.dart';
 class RequestUtils {
   const RequestUtils._();
 
-  static Future<T> getOrRetrieveSingle<T extends RestrrEntity>(
-      {required Id key,
-      required EntityCacheView<T> cacheView,
+  static Future<List<E>> fetchAllPaginated<E extends RestrrEntity<E, ID>, ID extends EntityId<E>>(
+      Restrr api, Paginated<E> firstBatch,
+      {Duration? delay}) async {
+    final List<E> all = [...firstBatch.items];
+    Paginated<E> current = firstBatch;
+    while (current.hasNext) {
+      if (delay != null) {
+        await Future.delayed(delay);
+      }
+      final Paginated<E> next = await current.nextPage!.call(api);
+      current = next;
+      all.addAll(next.items);
+    }
+    return all;
+  }
+
+  static Future<bool> deleteSingle<E extends RestrrEntity<E, ID>, ID extends EntityId<E>>(
+      {required CompiledRoute compiledRoute,
+      required Restrr api,
+      required EntityId key,
+      required EntityCacheView<E, ID> cacheView,
+      bool noAuth = false}) async {
+    final RestResponse<bool> response = await RequestHandler.noResponseRequest(
+        route: compiledRoute, routeOptions: api.routeOptions, bearerToken: noAuth ? null : api.session.token);
+    if (response.hasData && response.data!) {
+      cacheView.remove(key.value);
+      return true;
+    }
+    return false;
+  }
+
+  static Future<E> getOrRetrieveSingle<E extends RestrrEntity<E, ID>, ID extends EntityId<E>>(
+      {required EntityId key,
+      required EntityCacheView<E, ID> cacheView,
       required CompiledRoute compiledRoute,
-      required T Function(dynamic) mapper,
+      required E Function(dynamic) mapper,
       bool forceRetrieve = false,
       bool noAuth = false}) async {
-    if (!forceRetrieve && cacheView.contains(key)) {
-      return cacheView.get(key)!;
+    if (!forceRetrieve && cacheView.contains(key.value)) {
+      return cacheView.get(key.value)!;
     }
-    final RestResponse<T> response = await RequestHandler.request(
+    final RestResponse<E> response = await RequestHandler.request(
         route: compiledRoute,
         routeOptions: cacheView.api.routeOptions,
         bearerToken: noAuth ? null : cacheView.api.session.token,
@@ -29,16 +60,16 @@ class RequestUtils {
     return response.data!;
   }
 
-  static Future<List<T>> getOrRetrieveMulti<T extends RestrrEntity>(
-      {required BatchCacheView<T> batchCache,
+  static Future<List<E>> getOrRetrieveMulti<E extends RestrrEntity<E, ID>, ID extends EntityId<E>>(
+      {required BatchCacheView<E, ID> batchCache,
       required CompiledRoute compiledRoute,
-      required T Function(dynamic) mapper,
+      required E Function(dynamic) mapper,
       bool forceRetrieve = false,
       bool noAuth = false}) async {
     if (!forceRetrieve && batchCache.hasSnapshot) {
       return batchCache.get()!;
     }
-    final RestResponse<List<T>> response = await RequestHandler.multiRequest(
+    final RestResponse<List<E>> response = await RequestHandler.multiRequest(
         route: compiledRoute,
         routeOptions: batchCache.api.routeOptions,
         bearerToken: noAuth ? null : batchCache.api.session.token,
@@ -46,15 +77,15 @@ class RequestUtils {
     if (response.hasError) {
       throw response.error!;
     }
-    final List<T> remote = response.data!;
+    final List<E> remote = response.data!;
     batchCache.update(remote);
     return remote;
   }
 
-  static Future<Paginated<T>> getOrRetrievePage<T extends RestrrEntity>(
-      {required PageCacheView<T> pageCache,
+  static Future<Paginated<E>> getOrRetrievePage<E extends RestrrEntity<E, ID>, ID extends EntityId<E>>(
+      {required PageCacheView<E, ID> pageCache,
       required CompiledRoute compiledRoute,
-      required T Function(dynamic) mapper,
+      required E Function(dynamic) mapper,
       required int page,
       required int limit,
       bool forceRetrieve = false,
@@ -62,7 +93,7 @@ class RequestUtils {
     if (!forceRetrieve && pageCache.contains((page, limit))) {
       return pageCache.get((page, limit))!;
     }
-    final RestResponse<List<T>> response = await RequestHandler.paginatedRequest(
+    final RestResponse<List<E>> response = await RequestHandler.paginatedRequest(
         route: compiledRoute,
         routeOptions: pageCache.api.routeOptions,
         page: page,
@@ -72,7 +103,7 @@ class RequestUtils {
     if (response.hasError) {
       throw response.error!;
     }
-    final Paginated<T> remote = (response as PaginatedResponse<T>).toPage();
+    final Paginated<E> remote = (response as PaginatedResponse<E>).toPage();
     pageCache.cache(remote);
     return remote;
   }